chore: rename cert packages with plural prefix
