--- pkgs/community/swarmauri_cert_acme/README.md
@@ -1,20 +1,0 @@
-![Swamauri Logo](https://res.cloudinary.com/dbjmpekvl/image/upload/v1730099724/Swarmauri-logo-lockup-2048x757_hww01w.png)
-
-<p align="center">
-    <a href="https://pypi.org/project/swarmauri_cert_acme/">
-        <img src="https://img.shields.io/pypi/dm/swarmauri_cert_acme" alt="PyPI - Downloads"/></a>
-    <a href="https://hits.sh/github.com/swarmauri/swarmauri-sdk/tree/master/pkgs/community/swarmauri_cert_acme/">
-        <img alt="Hits" src="https://hits.sh/github.com/swarmauri/swarmauri-sdk/tree/master/pkgs/community/swarmauri_cert_acme.svg"/></a>
-    <a href="https://pypi.org/project/swarmauri_cert_acme/">
-        <img src="https://img.shields.io/pypi/pyversions/swarmauri_cert_acme" alt="PyPI - Python Version"/></a>
-    <a href="https://pypi.org/project/swarmauri_cert_acme/">
-        <img src="https://img.shields.io/pypi/l/swarmauri_cert_acme" alt="PyPI - License"/></a>
-    <a href="https://pypi.org/project/swarmauri_cert_acme/">
-        <img src="https://img.shields.io/pypi/v/swarmauri_cert_acme?label=swarmauri_cert_acme&color=green" alt="PyPI - swarmauri_cert_acme"/></a>
-</p>
-
----
-
-# Swarmauri ACME Certificate Service
-
-Community plugin providing an ACME (RFC 8555) certificate service built on top of Swarmauri's certificate interfaces.
