--- pkgs/community/swarmauri_cert_acme/tests/functional/certs/test_acmecertservice_functional.py
@@ -1,55 +1,0 @@
-import datetime
-
-import pytest
-from cryptography import x509
-from cryptography.hazmat.primitives import hashes, serialization
-from cryptography.hazmat.primitives.asymmetric import rsa
-from cryptography.x509.oid import NameOID
-
-from swarmauri_cert_acme import AcmeCertService
-from swarmauri_core.certs.ICertService import SubjectSpec
-from swarmauri_core.crypto.types import ExportPolicy, KeyRef, KeyType, KeyUse
-
-
-@pytest.mark.acceptance
-@pytest.mark.asyncio
-async def test_create_csr_and_verify_cert() -> None:
-    key = rsa.generate_private_key(public_exponent=65537, key_size=2048)
-    priv_pem = key.private_bytes(
-        serialization.Encoding.PEM,
-        serialization.PrivateFormat.TraditionalOpenSSL,
-        serialization.NoEncryption(),
-    )
-    account = KeyRef(
-        kid="acc",
-        version=1,
-        type=KeyType.RSA,
-        uses=(KeyUse.SIGN,),
-        export_policy=ExportPolicy.SECRET_WHEN_ALLOWED,
-        material=priv_pem,
-    )
-    svc = AcmeCertService(account_key=account)
-
-    subject: SubjectSpec = {"CN": "example.com"}
-    csr = await svc.create_csr(account, subject)
-    x509.load_pem_x509_csr(csr)
-
-    builder = (
-        x509.CertificateBuilder()
-        .subject_name(
-            x509.Name([x509.NameAttribute(NameOID.COMMON_NAME, "example.com")])
-        )
-        .issuer_name(
-            x509.Name([x509.NameAttribute(NameOID.COMMON_NAME, "example.com")])
-        )
-        .public_key(key.public_key())
-        .serial_number(x509.random_serial_number())
-        .not_valid_before(datetime.datetime.utcnow() - datetime.timedelta(days=1))
-        .not_valid_after(datetime.datetime.utcnow() + datetime.timedelta(days=1))
-        .add_extension(x509.BasicConstraints(ca=False, path_length=None), critical=True)
-    )
-    cert = builder.sign(key, hashes.SHA256())
-    cert_pem = cert.public_bytes(serialization.Encoding.PEM)
-
-    result = await svc.verify_cert(cert_pem)
-    assert result["valid"]
