--- pkgs/community/swarmauri_cert_acme/tests/unit/certs/test_acmecertservice_rfc2986.py
@@ -1,8 +1,0 @@
-import pytest
-
-from swarmauri_cert_acme import AcmeCertService
-
-
-@pytest.mark.unit
-def test_acmecertservice_mentions_rfc2986() -> None:
-    assert "RFC 2986" in AcmeCertService.__doc__
