--- pkgs/community/swarmauri_cert_cfssl/README.md
@@ -1,22 +1,0 @@
-
-![Swamauri Logo](https://res.cloudinary.com/dbjmpekvl/image/upload/v1730099724/Swarmauri-logo-lockup-2048x757_hww01w.png)
-
-<p align="center">
-    <a href="https://pypi.org/project/swarmauri_cert_cfssl/">
-        <img src="https://img.shields.io/pypi/dm/swarmauri_cert_cfssl" alt="PyPI - Downloads"/></a>
-    <a href="https://hits.sh/github.com/swarmauri/swarmauri-sdk/tree/master/pkgs/community/swarmauri_cert_cfssl/">
-        <img alt="Hits" src="https://hits.sh/github.com/swarmauri/swarmauri-sdk/tree/master/pkgs/community/swarmauri_cert_cfssl.svg"/></a>
-    <a href="https://pypi.org/project/swarmauri_cert_cfssl/">
-        <img src="https://img.shields.io/pypi/pyversions/swarmauri_cert_cfssl" alt="PyPI - Python Version"/></a>
-    <a href="https://pypi.org/project/swarmauri_cert_cfssl/">
-        <img src="https://img.shields.io/pypi/l/swarmauri_cert_cfssl" alt="PyPI - License"/></a>
-    <a href="https://pypi.org/project/swarmauri_cert_cfssl/">
-        <img src="https://img.shields.io/pypi/v/swarmauri_cert_cfssl?label=swarmauri_cert_cfssl&color=green" alt="PyPI - swarmauri_cert_cfssl"/></a>
-</p>
-
----
-
-# Swarmauri Cert Cfssl
-
-CFSSL-backed certificate service for Swarmauri.
-
