--- pkgs/community/swarmauri_cert_ocspverify/tests/perf/test_ocspverify_perf.py
@@ -1,10 +1,0 @@
-import pytest
-
-from swarmauri_cert_ocspverify import OcspVerifyService
-
-
-@pytest.mark.perf
-def test_supports_perf(benchmark) -> None:
-    svc = OcspVerifyService()
-    result = benchmark(svc.supports)
-    assert "ocsp" in result["features"]
