--- pkgs/community/swarmauri_cert_ocspverify/tests/unit/test_rfc5280.py
@@ -1,8 +1,0 @@
-import pytest
-
-from swarmauri_cert_ocspverify import OcspVerifyService
-
-
-@pytest.mark.unit
-def test_ocspverify_mentions_rfc5280() -> None:
-    assert "RFC 5280" in OcspVerifyService.__doc__
