--- pkgs/community/swarmauri_cert_ocspverify/tests/unit/test_rfc6960.py
@@ -1,8 +1,0 @@
-import pytest
-
-from swarmauri_cert_ocspverify import OcspVerifyService
-
-
-@pytest.mark.unit
-def test_ocspverify_mentions_rfc6960() -> None:
-    assert "RFC 6960" in OcspVerifyService.__doc__
