+++ pkgs/community/swarmauri_certs_acme/README.md
@@ -1,0 +1,20 @@
+![Swamauri Logo](https://res.cloudinary.com/dbjmpekvl/image/upload/v1730099724/Swarmauri-logo-lockup-2048x757_hww01w.png)
+
+<p align="center">
+    <a href="https://pypi.org/project/swarmauri_certs_acme/">
+        <img src="https://img.shields.io/pypi/dm/swarmauri_certs_acme" alt="PyPI - Downloads"/></a>
+    <a href="https://hits.sh/github.com/swarmauri/swarmauri-sdk/tree/master/pkgs/community/swarmauri_certs_acme/">
+        <img alt="Hits" src="https://hits.sh/github.com/swarmauri/swarmauri-sdk/tree/master/pkgs/community/swarmauri_certs_acme.svg"/></a>
+    <a href="https://pypi.org/project/swarmauri_certs_acme/">
+        <img src="https://img.shields.io/pypi/pyversions/swarmauri_certs_acme" alt="PyPI - Python Version"/></a>
+    <a href="https://pypi.org/project/swarmauri_certs_acme/">
+        <img src="https://img.shields.io/pypi/l/swarmauri_certs_acme" alt="PyPI - License"/></a>
+    <a href="https://pypi.org/project/swarmauri_certs_acme/">
+        <img src="https://img.shields.io/pypi/v/swarmauri_certs_acme?label=swarmauri_certs_acme&color=green" alt="PyPI - swarmauri_certs_acme"/></a>
+</p>
+
+---
+
+# Swarmauri ACME Certificate Service
+
+Community plugin providing an ACME (RFC 8555) certificate service built on top of Swarmauri's certificate interfaces.
