+++ pkgs/community/swarmauri_certs_cfssl/README.md
@@ -1,0 +1,22 @@
+
+![Swamauri Logo](https://res.cloudinary.com/dbjmpekvl/image/upload/v1730099724/Swarmauri-logo-lockup-2048x757_hww01w.png)
+
+<p align="center">
+    <a href="https://pypi.org/project/swarmauri_certs_cfssl/">
+        <img src="https://img.shields.io/pypi/dm/swarmauri_certs_cfssl" alt="PyPI - Downloads"/></a>
+    <a href="https://hits.sh/github.com/swarmauri/swarmauri-sdk/tree/master/pkgs/community/swarmauri_certs_cfssl/">
+        <img alt="Hits" src="https://hits.sh/github.com/swarmauri/swarmauri-sdk/tree/master/pkgs/community/swarmauri_certs_cfssl.svg"/></a>
+    <a href="https://pypi.org/project/swarmauri_certs_cfssl/">
+        <img src="https://img.shields.io/pypi/pyversions/swarmauri_certs_cfssl" alt="PyPI - Python Version"/></a>
+    <a href="https://pypi.org/project/swarmauri_certs_cfssl/">
+        <img src="https://img.shields.io/pypi/l/swarmauri_certs_cfssl" alt="PyPI - License"/></a>
+    <a href="https://pypi.org/project/swarmauri_certs_cfssl/">
+        <img src="https://img.shields.io/pypi/v/swarmauri_certs_cfssl?label=swarmauri_certs_cfssl&color=green" alt="PyPI - swarmauri_certs_cfssl"/></a>
+</p>
+
+---
+
+# Swarmauri Cert Cfssl
+
+CFSSL-backed certificate service for Swarmauri.
+
